Fix async for Flask 2.0
