+++ apiflask/app.py
@@ -320,7 +320,9 @@ class APIFlask(Flask):
         ):
             return self.make_default_options_response()  # pragma: no cover
         # otherwise dispatch to the handler for that endpoint
-        return self.view_functions[rule.endpoint](*req.view_args.values())  # type: ignore
+        return self.ensure_sync(  # type: ignore
+            self.view_functions[rule.endpoint]
+        )(*req.view_args.values())
 
     def error_processor(
         self,
