+++ tests/test_decorator_output.py
@@ -1,4 +1,3 @@
-import pytest
 from flask import make_response
 from flask.views import MethodView
 from openapi_spec_validator import validate_spec
@@ -239,25 +238,3 @@ def test_output_response_object_directly(app, client):
     rv = client.get('/foo')
     assert rv.status_code == 200
     assert rv.json['message'] == 'hello'
-
-
-def test_output_on_async_view(app, client):
-    if not hasattr(app, 'ensure_sync'):
-        pytest.skip('This test requires Flask 2.0 or higher')
-
-    @app.get('/foo')
-    @output(FooSchema)
-    async def foo():
-        return {'id': 1, 'name': 'foo'}
-
-    rv = client.get('/foo')
-    assert rv.status_code == 200
-    assert rv.json['id'] == 1
-    assert rv.json['name'] == 'foo'
-
-    rv = client.get('/openapi.json')
-    assert rv.status_code == 200
-    validate_spec(rv.json)
-    assert rv.json['paths']['/foo']['get']['responses']['200']
-    assert rv.json['paths']['/foo']['get']['responses']['200'][
-        'content']['application/json']['schema']['$ref'] == '#/components/schemas/Foo'
